Bug#11752731 - Innodb compiler warnings on windows 64-bit
Eliminate the hundreds of compiler warnings when compiling InnoDB
of Windows 64-bit.

+++ storage/innobase/trx/trx0trx.cc
@@ -2284,7 +2284,7 @@ state_ok:
 	}
 
 	if (trx->mysql_thd != NULL) {
-		innobase_mysql_print_thd(f, trx->mysql_thd, max_query_len);
+		innobase_mysql_print_thd(f, trx->mysql_thd, (uint) max_query_len);
 	}
 }
